pyproject.toml
[tool.poetry]
name = "bkg-rm-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
black = "^23.9.1"
pillow = "^10.0.1"


[tool.poetry.group.runpod.dependencies]
runpod = "^1.2.1"
onnxruntime = "1.15.1"
torchvision = "^0.15.2"
opencv-python = "^4.8.0.76"


[tool.poetry.group.local.dependencies]
onnxruntime-silicon = "^1.15.0"


[tool.poetry.group.api.dependencies]
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.21"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
boto3 = "^1.28.59"
pydantic-settings = "^2.0.3"
requests = "^2.31.0"
alembic = "^1.12.0"
psycopg2 = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
